refactor: moving meeting detail files to MeetingsPage/SalesmenMeeting folder and removing salesmen folder
