--- src/pages/salesman/MeetingsManagement/MeetingDetail/MeetingDetail.interface.ts
@@ -1,4 +1,0 @@
-import type { TMeetingContextMetadata } from '@services/models/MeetingSchema';
-
-export type TMeetingDetail = TMeetingContextMetadata;
-export type TMeetingTab = 'context' | 'insights' | 'action-plan';
