+++ src/router.tsx
@@ -7,9 +7,9 @@ import { Login } from '@pages/Login/Login';
 import { ManagersDetails } from '@pages/ManagersPage/ManagersDetails/ManagersDetails';
 import { ManagersPage } from '@pages/ManagersPage/ManagersPage';
 import { MeetingsPage } from '@pages/MeetingsPage/MeetingsPage';
+import { MeetingDetail } from '@pages/MeetingsPage/SalesmanMeetingsPage/SalesmanMeetingDetail/MeetingDetail';
 import { PageNotFound } from '@pages/PageNotFound/PageNotFound';
 import { RootComponent } from '@pages/RootComponent/RootComponent';
-import { MeetingDetail } from '@pages/salesman/MeetingsManagement/MeetingDetail/MeetingDetail';
 import { SalesmanDetail } from '@pages/SalesmenPage/SalesmanDetail/SalesmanDetail';
 import { SalesmenPage } from '@pages/SalesmenPage/SalesmenPage';
 import { useAuthStore } from '@store/authStore';
